requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
PyGithub==2.4.0
google-generativeai==0.8.3
anthropic==0.34.2
httpx==0.27.2
semgrep==1.90.0
